+++ biome.json
@@ -11,6 +11,7 @@
       "**/package.json",
       "**/packages/**/src/*",
       "**/examples/**/src/*",
+      "**/examples/**/*.json",
       "**/examples/**/webpack.config.js",
       "./packages/pack/docs/features-list.json",
       "crates/pack-core/js/**/*.ts",
