+++ examples/utooweb-demo/src/raw/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+
+console.log(React, createRoot)
+// import Demo from './App';
+//
+// createRoot(document.getElementById('root')).render(<Demo />);
+
+
