+++ examples/utooweb-demo/src/raw/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "muse-browser-example-antd-enhance",
+  "dependencies": {
+    "@ant-design/icons": "^5.2.0",
+    "antd": "latest",
+    "dayjs": "^1.11.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
+  "devDependencies": {
+    "@types/react": "^18.2.67",
+    "@types/react-dom": "^18.2.22"
+  }
+}
